Create Caesar Cipher
A Caesar cipher, also known as  shift cipher, Caesar's code or Caesar shift, is one of the simplest and most widely known encryption technique. It is a type of substitution cipher in which each letter in the plaintext is replaced by a letter some fixed number of positions down the alphabet.