Create Monoalphabetic Cipher
A mono-alphabetic cipher  is a substitution cipher where each letter of the plain text is replaced with another letter of the alphabet. It uses a fixed key which consist of the 26 letters of a “shuffled alphabet” . This type of cipher is a form of symmetric encryption as the same key is used to both encrypt and decrypt a message.