Create  POLY SUBSTITUTION CIPHER(ONE TIME PAD)
One Time Pad Cipher is the improvement of the Vernam Cipher, which is a poly substitution cipher. It is the only available cipher that is unbreakable(completely secure). It is a method of encrypting alphabetic plain text. It is one of the Substitution techniques which converts plain text into ciphertext. In this mechanism, we assign a number to each character of the Plain-Text.

The two requirements for the One-Time pad are

1 The key should be randomly generated as long as the size of the message.
2 The key is to be used to encrypt and decrypt a single message, and then it is discarded.